# perf: 优化小除数除法路径