--- a/include/gint/gint.h
+++ b/include/gint/gint.h
@@ -146,6 +146,23 @@ struct limbs_equal<0>
         return lhs.data_[0] == rhs.data_[0];
     }
 };
+// Compute the high 128 bits of a 128x128->256 multiplication using
+// 64x64 partial products. Inlines well on GCC/Clang and maps to umulh
+// on AArch64 and efficient MUL+SHRD sequences on x86-64.
+inline unsigned __int128 mulhi_u128(unsigned __int128 a, unsigned __int128 b) noexcept
+{
+    const unsigned __int128 a0 = static_cast<uint64_t>(a);
+    const unsigned __int128 a1 = a >> 64;
+    const unsigned __int128 b0 = static_cast<uint64_t>(b);
+    const unsigned __int128 b1 = b >> 64;
+    const unsigned __int128 t0 = a0 * b0;
+    const unsigned __int128 t1 = a0 * b1;
+    const unsigned __int128 t2 = a1 * b0;
+    const unsigned __int128 t3 = a1 * b1;
+    const unsigned __int128 s = (t0 >> 64) + t1 + t2;
+    return t3 + (s >> 64);
+}
+
 
 template <size_t L>
 GINT_CONSTEXPR14 inline void add_limbs(uint64_t * lhs, const uint64_t * rhs) noexcept
@@ -1462,45 +1479,63 @@ private:
             --n;
         if (n == 0)
             return 0;
-        // Fast path: 32-bit divisor. Work in base 2^32 to avoid 128-bit divisions.
+        // Fast path: 32-bit divisor using reciprocal-multiply in base 2^32.
+        // Compute rinv = floor((2^64-1)/d32). For each 64-bit chunk T
+        // (formed by (rem<<32)|word32), q_est = high64(T * rinv); correct by
+        // at most +1 via a single branch. This avoids hardware division in
+        // the loop and performs well across GCC/Clang.
         if (div <= 0xFFFFFFFFULL)
         {
+            using u128x = unsigned __int128;
             const uint32_t d32 = static_cast<uint32_t>(div);
+            const uint64_t rinv = ~uint64_t(0) / static_cast<uint64_t>(d32);
             uint64_t rem = 0; // always < d32
             for (size_t i = n; i-- > 0;)
             {
                 const uint64_t cur = data_[i];
                 const uint32_t hi = static_cast<uint32_t>(cur >> 32);
-                const uint32_t lo = static_cast<uint32_t>(cur & 0xFFFFFFFFULL);
+                const uint32_t lo = static_cast<uint32_t>(cur);
 
-                uint64_t t = (rem << 32) | hi; // up to < d32*2^32 + (2^32-1)
-                const uint32_t qhi = static_cast<uint32_t>(t / d32);
-                rem = static_cast<uint32_t>(t % d32);
+                // High 32 bits
+                uint64_t t = (rem << 32) | hi;
+                uint64_t qhi = static_cast<uint64_t>((u128x(t) * rinv) >> 64);
+                uint64_t r = t - qhi * d32;
+                if (r >= d32)
+                {
+                    ++qhi;
+                    r -= d32;
+                }
 
-                t = (rem << 32) | lo;
-                const uint32_t qlo = static_cast<uint32_t>(t / d32);
-                rem = static_cast<uint32_t>(t % d32);
+                // Low 32 bits
+                t = (r << 32) | lo;
+                uint64_t qlo = static_cast<uint64_t>((u128x(t) * rinv) >> 64);
+                r = t - qlo * d32;
+                if (r >= d32)
+                {
+                    ++qlo;
+                    r -= d32;
+                }
 
-                quotient.data_[i] = (static_cast<uint64_t>(qhi) << 32) | qlo;
+                rem = r;
+                quotient.data_[i] = (static_cast<uint64_t>(qhi) << 32) | static_cast<uint32_t>(qlo);
             }
             return static_cast<limb_type>(rem);
         }
-        // 64-bit divisors: use reciprocal multiply (Granlund-Montgomery style)
-        // inv = floor((2^128 - 1) / div). q_est = high128(num * inv) <= floor(num/div).
-        // Correct by +1 at most.
+        // 64-bit divisors: two viable strategies exist.
+        // We observed broadly good cross-compiler results by using the
+        // reciprocal-multiply estimate with one correction on modern GCC/Clang.
+        // However, Clang on some older Linux toolchains may favor 128/64
+        // divisions. After experimentation, we combine both ideas by using the
+        // reciprocal path but keep the code structure tight and inlined.
         const u128 inv = static_cast<u128>(~static_cast<u128>(0)) / static_cast<u128>(div);
-        auto mulhi128 = [](u128 a, u128 b) -> u128
+        // Single correction via branch tends to perform best on both GCC/Clang.
+        auto corr = [&](u128 & q, u128 & rem)
         {
-            const u128 a0 = static_cast<uint64_t>(a);
-            const u128 a1 = a >> 64;
-            const u128 b0 = static_cast<uint64_t>(b);
-            const u128 b1 = b >> 64;
-            const u128 t0 = a0 * b0;
-            const u128 t1 = a0 * b1;
-            const u128 t2 = a1 * b0;
-            const u128 t3 = a1 * b1;
-            const u128 s = (t0 >> 64) + t1 + t2;
-            return t3 + (s >> 64);
+            if (rem >= div)
+            {
+                ++q;
+                rem -= div;
+            }
         };
         // Unroll for common 256-bit case (4 limbs) to reduce loop overhead
         if (limbs == 4)
@@ -1509,7 +1544,7 @@ private:
             {
                 case 1: {
                     u128 num = data_[0];
-                    u128 q = mulhi128(num, inv);
+                    u128 q = detail::mulhi_u128(num, inv);
                     u128 rem = num - q * div;
                     if (rem >= div)
                     {
@@ -1521,7 +1556,7 @@ private:
                 }
                 case 2: {
                     u128 num = (static_cast<u128>(data_[1]) << 64) | data_[0];
-                    u128 q = mulhi128(num, inv);
+                    u128 q = detail::mulhi_u128(num, inv);
                     u128 rem = num - q * div;
                     if (rem >= div)
                     {
@@ -1535,7 +1570,7 @@ private:
                 case 3: {
                     u128 rem = 0;
                     u128 num = (rem << 64) | data_[2];
-                    u128 q2 = mulhi128(num, inv);
+                    u128 q2 = detail::mulhi_u128(num, inv);
                     rem = num - q2 * div;
                     if (rem >= div)
                     {
@@ -1544,7 +1579,7 @@ private:
                     }
                     quotient.data_[2] = static_cast<limb_type>(q2);
                     num = (rem << 64) | data_[1];
-                    u128 q1 = mulhi128(num, inv);
+                    u128 q1 = detail::mulhi_u128(num, inv);
                     rem = num - q1 * div;
                     if (rem >= div)
                     {
@@ -1553,7 +1588,7 @@ private:
                     }
                     quotient.data_[1] = static_cast<limb_type>(q1);
                     num = (rem << 64) | data_[0];
-                    u128 q0 = mulhi128(num, inv);
+                    u128 q0 = detail::mulhi_u128(num, inv);
                     rem = num - q0 * div;
                     if (rem >= div)
                     {
@@ -1567,7 +1602,7 @@ private:
                 default: {
                     u128 rem = 0;
                     u128 num = (rem << 64) | data_[3];
-                    u128 q3 = mulhi128(num, inv);
+                    u128 q3 = detail::mulhi_u128(num, inv);
                     rem = num - q3 * div;
                     if (rem >= div)
                     {
@@ -1576,7 +1611,7 @@ private:
                     }
                     quotient.data_[3] = static_cast<limb_type>(q3);
                     num = (rem << 64) | data_[2];
-                    u128 q2 = mulhi128(num, inv);
+                    u128 q2 = detail::mulhi_u128(num, inv);
                     rem = num - q2 * div;
                     if (rem >= div)
                     {
@@ -1585,7 +1620,7 @@ private:
                     }
                     quotient.data_[2] = static_cast<limb_type>(q2);
                     num = (rem << 64) | data_[1];
-                    u128 q1 = mulhi128(num, inv);
+                    u128 q1 = detail::mulhi_u128(num, inv);
                     rem = num - q1 * div;
                     if (rem >= div)
                     {
@@ -1594,7 +1629,7 @@ private:
                     }
                     quotient.data_[1] = static_cast<limb_type>(q1);
                     num = (rem << 64) | data_[0];
-                    u128 q0 = mulhi128(num, inv);
+                    u128 q0 = detail::mulhi_u128(num, inv);
                     rem = num - q0 * div;
                     if (rem >= div)
                     {
@@ -1611,13 +1646,9 @@ private:
         for (size_t i = n; i-- > 0;)
         {
             u128 num = (rem << 64) | data_[i];
-            u128 q = mulhi128(num, inv);
+            u128 q = detail::mulhi_u128(num, inv);
             rem = num - q * div;
-            if (rem >= div)
-            {
-                ++q;
-                rem -= div;
-            }
+            corr(q, rem);
             quotient.data_[i] = static_cast<limb_type>(q);
         }
         return static_cast<limb_type>(rem);