标题：perf: 优化小除数除法路径

动机
- 目标场景：针对 256 位整数的小除数除法性能瓶颈
- 基线：BM_Division<256> 9.61 ns

优化思路
- 算法/路径：引入 mulhi_u128 以及倒数乘法估算，消除循环中的硬件除法
- 位宽专用化：为常见 256 位路径展开
- 可维护性：代码结构保持紧凑

正确性
- 语义不变：保持严格位宽与操作语义
- 覆盖测试：`make test` 通过

基准对比（必填）
- 环境：5×2.79GHz CPU, GCC, Debug benchmark 库
- 方法：`make bench BENCH_ARGS="--benchmark_filter=BM_Division --benchmark_min_time=0.1s"`
- 数据：变更前 BM_Division<256> 9.61 ns；变更后 9.31 ns（≈‑3.1%）
- 总结：小除数除法有小幅提升，其他场景变化不显著

回归与风险
- 平台差异：依赖 `__int128`，不同编译器可能对倒数乘法优化差异
- 代码尺寸与可读性：新增辅助函数，复杂度可控

清单（提交前请逐项勾选）
- [x] 分支名为英文（PR 指南要求）
- [x] 已运行 clang-format（使用仓库根目录 `.clang-format`）
- [x] 新增/更新了 `tests/` 覆盖关键路径，`make test` 通过
- [x] 已提供“变更前/后”基准数据（必填）
- [ ] 若更改基准用例，已更新 `bench/` 与 `docs/BENCHMARKS.md`
- [x] 保持 C++11 兼容与 Header-only 特性
- [x] 保持严格位宽与原生类型互操作语义不变


------
https://chatgpt.com/codex/tasks/task_e_68b57a17be3c8329b28a2b0af9c3cc7a